Shift `hoistImportDeclaration` to transpiler
